backend/pyproject.toml
[project]
name = "kakeibo-backend"
version = "0.1.0"
description = "FastAPI backend for the kakeibo app."
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.13.0",
    "fastapi>=0.115.0",
    "pydantic-settings>=2.6.0",
    "pyjwt>=2.9.0",
    "pymupdf>=1.24.0",
    "pymysql>=1.1.0",
    "sqlalchemy>=2.0.0",
    "uvicorn[standard]>=0.32.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "-q"

